refactor: 구매 모달 창의 handler를 수동 구매와 자동 구매에 대해 분리
