--- src/js/Model/Ticket.js
@@ -1,23 +1,0 @@
-export default class Ticket {
-  constructor() {
-    this.numbers = [];
-    this.winningRank = 0;
-    this.profit = 0;
-  }
-
-  setNumbers(numbers) {
-    this.numbers = numbers;
-  }
-
-  getNumbers() {
-    return this.numbers;
-  }
-
-  setWinningRank(winningRank) {
-    this.winningRank = winningRank;
-  }
-
-  setProfit(profit) {
-    this.profit = profit;
-  }
-}
