+++ src/js/index.js
@@ -1,7 +1,8 @@
 import Lotto from './model/Lotto.js';
 import { $ } from './utils/querySelector.js';
 import { handlePurchasePriceInput } from './handler/handlePurchasePriceInput.js';
-import { handlePurchaseModal } from './handler/handlePurchaseModal.js';
+import { handleSelfNumberInput } from './handler/handleSelfNumberInput.js';
+import { handleAutoNumberInput } from './handler/handleAutoNumberInput.js';
 import { handlePurchaseResultToggle } from './handler/handlePurchaseResultToggle.js';
 import { handleWinningNumberInput } from './handler/handleWinningNumberInput.js';
 import { handleResultModal } from './handler/handleResultModal.js';
@@ -14,9 +15,14 @@ const init = () => {
     handlePurchasePriceInput(lotto);
   });
 
-  $('#purchase-modal').addEventListener('submit', (event) => {
+  $('#purchase-modal__self-input-form').addEventListener('submit', (event) => {
     event.preventDefault();
-    handlePurchaseModal(event, lotto);
+    handleSelfNumberInput(lotto);
+  });
+
+  $('#purchase-modal__auto-input-form').addEventListener('submit', (event) => {
+    event.preventDefault();
+    handleAutoNumberInput(lotto);
   });
 
   $('#purchase-result-section__toggle').addEventListener(
